requirements.txt
Flask==2.0.1
Pillow==8.2.0
tensorflow==2.6.0